Use model meta to generate verbose name
Render field names of object rather than the word "None"

This is fix for issue #53, since an object may not exist when rendering
`templates/admin/change_data.html` - ie, if we're greating a new object
when `confirm_add = True` is set on a `ModelAdmin` with
the `AdminConfirmMixin`.

Also makes sure we send an actual object (not the string "None") to the
base template, to remove other instances of the word "None" from
the add view.